Moved rules getter to separate file, matching other commands
