+++ commands/compliance/rules/rules.go
@@ -1,7 +1,6 @@
 package rules
 
 import (
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/spf13/cobra"
@@ -78,22 +77,6 @@ Available Rules:
 		},
 	}
 )
-
-var GetterCommands = []*cobra.Command{
-	&cobra.Command{
-		Use:     "name <address>",
-		Short:   "Gets the name of the rule at the <address>",
-		Example: "t0ken rules name 0xa16e2faddd9469e5076c33ae53a2c6f40a058413",
-		Args:    cli.AddressArgFunc("address", 0),
-		Run: func(cmd *cobra.Command, args []string) {
-			addr := common.HexToAddress(args[0])
-			cli.Connect(cmd, args)
-			caller, err := rules.NewRuleCaller(addr, cli.Conn.Client)
-			cli.CheckErr(cmd, err)
-			cli.CheckGetter(cmd)(caller.Name(&bind.CallOpts{}))
-		},
-	},
-}
 
 func init() {
 	// Add the 'gasPrice', 'nonce' and 'wait' flags to the deploy function
